Move snapping logic into SnapFinder
